Fix adding cloud-init logs to diagnostics archive (#3710)
* Fix adding cloud-init logs to diagnostics archive

* Add debug message when logs not found

+++ src/Agent.Worker/DiagnosticLogManager.cs
@@ -146,8 +146,16 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
                 string resultLogs = await DumpCloudInitLogs(logsFilePath);
                 executionContext.Debug(resultLogs);
 
-                string destination = Path.Combine(supportFilesFolder, Path.GetFileName(logsFilePath));
-                File.Copy(logsFilePath, destination);
+                if (File.Exists(logsFilePath))
+                {
+                    string destination = Path.Combine(supportFilesFolder, Path.GetFileName(logsFilePath));
+                    File.Copy(logsFilePath, destination);
+                    executionContext.Debug("Cloud-init logs added to the diagnostics archive.");
+                }
+                else
+                {
+                    executionContext.Debug("Cloud-init logs were not found.");
+                }
 
                 executionContext.Debug("Dumping cloud-init logs is ended.");
             }
